# Optimize `length+`, `dotted-list?` and `iota` in SRFI-1