1. Optimize `length+`, `dotted-list?`

Since we already have the C function `list_type_and_length()` in STklos core, it makes no sense to keep these in Scheme.

2. Optimize `iota`

Because it's easy. And we can allocate the full list at once now. The code in SRFI-1 will do N allocations...

BUT

We had to declare `list_type_and_length` non-static and include it in `stklos.h`...

Results:

```scheme
(time
  (repeat 10 (iota 10000000)))
New code:
1463.77 ms
2400000000 bytes in 10 allocation calls

Old code:
7940.318 ms
2400001920 bytes in 100000050 allocation calls

(time
  (let ((x (iota 1000000)))
    (repeat 100 (length+ x))))
New code:
507.476 ms
24000000 bytes in 1 allocation call

Old code:
5892.6 ms
24011392 bytes in 1000305 allocation calls

(let ((x (iota 1000000)))
  (set-cdr! (last-pair x) 1)
  (time (repeat 100 (length+ x))))
New code:
494.158 ms
Allocations: 0

Old code:
5714.828 ms
11200 bytes in 300 allocation calls
```